fix(deps): upgraded github.com/go-resty/resty/v2 v2.16.2 => v2.16.3
https://github.com/go-resty/resty/releases/tag/v2.16.3
