+++ compress/compress_test.go
